create working version, move filter function to services
